requirements.txt
pytz~=2023.3.post1
requests~=2.31.0
google-auth
google-api-python-client
pyyaml
pydantic